agents/the-oracle/pyproject.toml
[tool.poetry]
name = "the-oracle"
version = "0.1.0"
description = "Marketplace Intelligence Engine — rates, ranks, and recommends hackathon agent services"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
payments-py = {version = ">=1.1.0", extras = ["mcp"]}
fastapi = "^0.120.0"
uvicorn = ">=0.34.2,<1.0.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
